requirements.txt
# Core dependencies for web crawling and AI
crawl4ai[all]>=0.6.0
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0

# AI/ML dependencies
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.7.0

# Utility dependencies
python-dotenv>=0.19.0
pyyaml>=6.0
tenacity>=8.0.0
tqdm>=4.62.0
psutil>=5.8.0

# Web automation
playwright>=1.20.0
selenium>=4.0.0

# Optional but useful
colorama>=0.4.4
rich>=10.0.0 